# app/Resources/swagger.yml
swagger: '2.0'
info:
  title: Cookway - the way to cook
  description: Web application with recipes
  version: 1.0.0
  contact:
    email: mkuncicka@gmail.com
basePath: /api
...
schemes:
  - http
  - https
paths:
...
  /recipes:
    get:
      summary: List of recipes
      operationId: RecipeListQuery
      tags:
        - Recipes
      responses:
        '200':
          description: List of all recipes correctly returned
          schema:
            $ref: '#/definitions/RecipeList'
        '403':
          description: Request failed
    post:
      summary: Intention of creating a new recipe - adds new recipe to the database
      operationId: NewRecipeHandler
      tags:
        - Recipes
      parameters:
        - name: content
          in: body
          schema:
            $ref: '#/definitions/CreateRecipe'
          required: true
      responses:
        '200':
          description: Recipe succesfuly created
...
definitions:
...
  RecipeList:
    title: List of recipes
    description: List of recipes
    required:
      - data
      - totalCount
    properties:
      totalCount:
        title: Total count of recipes
        type: integer
      data:
        title: List of recipes
        type: array
        items:
          $ref: '#/definitions/RecipeListItemView'
  RecipeListItemView:
    title: Single recipe view model
    description: Single recipe view model presented on the list
    required:
      - title
      - ingredients
      - prescription
    properties:
      title:
        title: Title of the recipe
        type: string
      description:
        title: Description of the recipe
        type: string
      ingredients:
        title: List of ingredients
        type: array
        items:
          $ref: '#/definitions/IngredientListView'
      prescription:
        title: Prescription
        type: string
      preparationTime:
        title: Time of preparation in minutes
        type: integer
      preparationTimeText:
        title: Additional text to preparation time
        description: Text explaining what is included in preparation time
        type: string
      author:
        title: Author's name
        type: string
      photos:
        title: List of photos
        type: array
        items:
          $ref: '#/definitions/PhotoListView'
  PhotoListView:
    title: List of photos
    description: List of photos
    required:
      - data
      - totalCount
    properties:
      totalCount:
        title: Total count of photos
        type: integer
      data:
        title: List of photos
        type: array
        items:
          $ref: '#/definitions/PhotoListItemView'
  PhotoListItemView:
    title: Single photo view model
    description: Single photo connected with recipe
    required:
      - id
      - name
      - url
    properties:
      id:
        title: Photo's id
        type: integer
      name:
        title: Photo's name
        type: string
      url:
        title: Photo's displaying url
        type: string
      isMain:
        title: Information if this is a main recipe's photo
        type: boolean
  IngredientListView:
    title: List of ingredients
    description: List of ingredients
    required:
      - data
      - totalCount
    properties:
      totalCount:
        title: Total count of ingredients
        type: integer
      data:
        title: List of ingredients
        type: array
        items:
          $ref: '#/definitions/IngredientListItemView'
  IngredientListItemView:
    title: Single ingredient view model
    description: Single ingredient view model presented on the list
    required:
      - name
      - systemName
      - amount
      - unitName
    properties:
      name:
        title: Name of the ingredient
        type: string
      systemName:
        title: System name of the ingredient
        type: string
      amount:
        title: Amount of the ingredient
        type: number
      unitName:
        title: Unit of the ingredient amount
        type: string
  CreateRecipe:
    title: Recipe model view
    required:
      - title
      - prescription
    properties:
      title:
        title: Title of the recipe
        type: string
      ingredients:
        title: Ingredients from database
        type: array
        items:
          schema:
            $ref: '#/definitions/CreateIngredient'
      description:
        title: Description
        type: string
      prescription:
        title: Instructions how to make a meal
        type: string
      preparationTime:
        title: Time of preparation
        type: integer
      preparationTimeText:
        title: Additional text explaining time of preparation
        type: string
      photoId:
        title: Photo of the meal
        type: integer
  CreateIngredient:
    name:
      title: Nama of the ingredient
      type: string
    amount:
      title: Amount of the ingredient
      type: float
    unit:
      title: Unit of the amount
      type: number
...